requirements.txt
Django~=4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.0.1
defusedxml~=0.8.0rc2
cffi~=1.16.0
future~=0.18.2
pip~=23.2.1
wheel~=0.37.0
cryptography~=41.0.4
Jinja2~=3.1.2
pytz~=2023.3.post1
PyYAML~=6.0.1
setuptools~=68.0.0
asgiref~=3.7.2
sqlparse~=0.4.4
requests~=2.31.0
packaging~=23.1
coreschema~=0.0.4
inflection~=0.5.1
uritemplate~=4.1.1
coreapi~=2.3.3
MarkupSafe~=2.1.3
certifi~=2023.7.22
itypes~=1.2.0
idna~=3.4
urllib3~=2.0.6
PyJWT~=2.8.0